requirements.txt
cv2==1.0
docutils==0.12
future==0.15.2
graphviz==0.4.2
matplotlib==1.4.3
numpy==1.8.2
Pillow==2.8.2
python-telegram-bot==3.4
pytz==2013.7
scikit-image==0.11.3
scikit-learn==0.17.1
scipy===0.14.0.dev-5127d54
six==1.10.0
sklearn==0.0
virtualenv==13.0.3
virtualenv-clone==0.2.5
virtualenvwrapper==4.6.0